# Optimize AssetModel refresh time